http: fix ptr of conn and request at req_close
The request of the connection must be set to NULL at req_close.

Wrong checks in conn_idle would lead to not start the idle_timeout
and the request of the connection would not be set to NULL.
At TCP connection close, close_handler & try_next is called with
ECONNRESET. This would lead to a ptr deref in try_next of conn->req
which is valid data, but the req itself is not valid anymore.
conn_req is not a NULL ptr so the check if (!req) is not effective.
Following the next line, one sets a nonvalid data segment (req->conn) -> SEGV.