mk: do not overwrite LIBRE_PATH
A simpler way for cross compilation could be to set LIBRE_PATH externally. This
commit prevents that LIBRE_PATH is overwritten by re.mk and sets LIBRE_INC
and LIBRE_SO correctly by means of LIBRE_PATH. These variables are added to the
CFLAGS resp. to the "-L" argument of the linker command in baresip and rem
Makefile.