sdp: initialize SDP media dir to inactive
In baresip audio+video with SDP session and media is allocated by default for
each call. Thus it would expect video RTP packets, regardless if the remote
offers video or not. This fix achieves that the video remote direction stays at
inactive until an m-line for video is found in an SDP offer. Then the media
remote dir is initialized to the sessions media dir, which is send-receive. The
direction attribute may override to e.g. one-directional video.