sip: reuse transport protocol of first request in dialog (#143)
A SIP dialog which was initiated by a peer did not remember the transport
protocol. Following requests selected their transport protocol independently.
This lead in situations where the default protocol was selected differently on
two User-Agents to a protocol mix during a dialog. For UDP and TLS mix this
further lead to unprocessed BYE, re-INVITE, ...