jbuf: do not compute jitter for telev packets
Telephone-event packets also increment the RTP sequence number and thus have to
be put into the jitter buffer. Otherwise missing packets are detected wrongly.
But then the telev packets have to be ignored for the jitter computation. The
payload type of the first packet is used as reference to filter the audio/video
packets for the jitter computation.