set sdp laddr to SIP src address (#172)
* sip: add a pointer for the content to the send handler

This allows to postpone the encoding of the content with body (e.g. SDP) until
the source address of the SIP transport is known.

* sipsess: add a function pointer for the content description

This allows the application to postpone the encoding of the content description
(e.g. the SDP) until the source address of the SIP transport is known.

sipsess_connect() uses this for the INVITE. Other SIP session types do not set
the new function pointer.

* sipsess: set sent_offer flag after SIP content was encoded