sip: add connection configuration per peer address
A new setter function `sip_conncfg_set()` can be used in a send handler to set
a connection configuration. For now it allows to specify a source port for
TCP connections.

Some Session Border Controller (SBC) or other SIP components require the
ability to set a static source port at a UAC for SIP TCP connections.