sipreg: split allocation from send and add a setter
Adds a setter for the source port if TCP/TLS is used as transport protocol. In
order to allow changing properties of a SIP registration client (sipreg) the
allocation and sending has to be separated. The function `sipreg_register()` is
marked as deprecated and should be replaced by new `sipreg_alloc()` and
`sipreg_send()` in baresip and other applications.