sip: add optional TCP source port (#198)
* sip: add connection configuration per peer address

A new setter function `sip_conncfg_set()` can be used in a send handler to set
a connection configuration. For now it allows to specify a source port for
TCP connections.

Some Session Border Controller (SBC) or other SIP components require the
ability to set a static source port at a UAC for SIP TCP connections.

* sipreg: split allocation from send and add a setter

Adds a setter for the source port if TCP/TLS is used as transport protocol. In
order to allow changing properties of a SIP registration client (sipreg) the
allocation and sending has to be separated. The function `sipreg_register()` is
marked as deprecated and should be replaced by new `sipreg_alloc()` and
`sipreg_send()` in baresip and other applications.

* sipreg: fix clang warning struct contains uninitialized data

* sipreg: sipreg_set_srcport() does not need a return value