list: O(1) sorted insert if we expect append in most cases (#318)
Append means putting to tail. When the loop for the sorted insert is started
at the tail, then for a rising sequence we need only one check with a "less"
operator to know that the element has to be appended to the list.

Like `list_sort()` also `list_insert_sorted()` takes a less-equal operator.
Because we don't want to change this API detail, we declare an internal
function `le_less` by means of the given less-equal operator.