main: Add support for external threads attaching/detaching re context. (#414)
This commit allows to create an event loop context and keep a reference
to it in the caller application. After that, this and other threads
are allowed to attach to and detach from this context as needed. Threads
are synchronized by calling re_thread_enter/leave. The re context can
be destroyed by calling mem_deref.

This is useful to allow threads that are not running event loop to
invoke methods on an event loop context that is created and managed by
another thread.