mem: Align data to natural alignment.
Ensure that the pointer returned by mem_alloc & co. is properly aligned
similar to malloc on a given platform. This is important e.g. for SIMD
processing, including libc string operations.

Also, fixed mem_realloc behavior when there are multiple references to
the memory buffer being reallocated. In this case, the implementation
now allocates a new buffer and copies the contents and decrements the
reference counter on the original memory buffer. This way the other
references for the old buffer remain valid.

Also, fixed formatting in mem_status.