main: Remove the limit on fd values, relax maxfd limit for epoll.
Implemented binary search for fd handler by fd. This allows to
remove the limitation on the fd values to be below the maxfd limit
for backends other than select(). For select() the limitation is still
in place, as it is imposed by FD_SET macros.

As a side effect, this allowed to remove the Windows-specific code
to lookup the fd event handler as the common code is used on all
platforms now.

Also, for epoll() backend the maxfds limit is relaxed. It no longer
limits the number of fds that can be managed by epoll() but rather the
number of events that can be received in one syscall. epoll_create()
is not using the number of fds since Linux kernel 2.6.8.