rtp: Add realtime timestamps to rtp_send and rtcp_sess_tx_rtp methods.
The realtime timestamps allow the application to specify the time point
that corresponds to the RTP timestamp on the media presentation timeline.
This can be called "capture time", for whatever definition of "capturing"
the application has.

This timestamp is further used to calculate the NTP/RTP timestamps that
are sent in RTCP SR and used on the receiver to synchronize different
media streams from the sender. Having the timestamp provided by the
application allows for a more precise correspondence between NTP and RTP
timestamps, even if the media is sent non-monotonically and out-of-order.