atomic: Implemented abstraction layer similar to C11 atomics.
This new abstraction layer defines a set of macros that resemble C11
atomics interface and are implemented in terms of C11 atomics or compiler
intrinsics, depending on the compiler.

Besides C11 atomics, gcc __atomic, and MSVC Interlocked intrinsics are
supported. MSVC implementation has a caveat that it has to convert its
arguments and returned values to the largest supported integer size,
unsigned __int64. This may cause warnings in the calling code where
this results in implicit conversions. These warnings can be worked around
with explicit casts.