ice: fix ICE_CAND_TYPE_RELAY base
RFC 5245 and 8445 defines:

---
Base: The transport address that an ICE agent sends from for a
particular candidate.  For host, server-reflexive, and peer-
reflexive candidates, the base is the same as the host candidate.
For relayed candidates, the base is the same as the relayed candidate
(i.e., the transport address used by the TURN server to send from).
---

This allows also a RELAY only policy (without any other candidates),
like defined with iceTransportPolicy (WebRTC).