dns/client: fix rrlv reference cache handling (#637)
It's not enough to reference rr for handle caching, since if the original query is derefed the rr->le_priv->list is derefed too. This can lead to heap-use-after-free crashes if list is accessed (like with list_unlink). This PR references the complete rrlv lists for caching.

Fixes baresip/retest#168