misc: fix glibc feature selection (#674)
The __USE_* macros are for glibc internal use only and must not
be changed.  The _*_SOURCE macros must be defined before the
first glibc header file is included.  If these rules are not
followed, glibc header changes often introduce implicit function
declarations or break the build in other ways.

The most straightforward solution is to define _GNU_SOURCE at the
CMake level.  The basename function is not used, and the sources
already deal with the strerror_r API variant for _GNU_SOURCE.
(An alternative approach could remove the strerror_r handling,
and build with -D_XOPEN_SOURCE -D_DEFAULT_SOURCE, with a definition
of _GNU_SOURCE at the start of src/tcp/tcp.c)