sip: add connection handler for TCP
The added connection handler is called when a TCP connection is about to be
established and there was already a socket opened. At this point in time there
is also a source TCP port which should be used in the Via header.

In request.c the connect_handler() encodes the Via header with correct TCP
source port.