tls/sni: do not enable client verification when SNI matching is done
Currently, whenever a TLS handshake message is received and SNI
certificate matching is done, TLS peer verification is also enabled.
This ignores settings like verify_server and verify_client in the tls
struct and is not needed. Therefore, the call to explicitly enable peer
verification as part of SNI is removed.